updated imports for renamed middleware folder
