+++ server/api/routes/auth.route.js
@@ -4,7 +4,7 @@ const {
     me
 } = require("../controllers/auth.controller.js")
 
-const { isLoggedIn } = require("../utils/middleware.util.js")
+const { isLoggedIn } = require("../middleware/middleware.util.js")
 
 const router = require("express").Router()
 module.exports = router
